MCMC settings class

Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.

SCAMCMC(object, ...)

# S4 method for missing
SCAMCMC(object, ...)

getADMBCallArgs(object, ...)

# S4 method for SCAMCMC
getADMBCallArgs(object, ...)

getN(object, ...)

# S4 method for SCAMCMC
getN(object, ...)

Arguments

object
a SCAMCMC object

...
extra arguments

Slot

mcmc N

Run N MCMC iterations

mcsave N

Save every N th MCMC iteration

mcscale N

Rescale step size for first N iterations

mcmult N

Rescale the covariance matrix

mcrb N

Reduce high parameter correlations

mcprobe X

Use a fat-tailed proposal distribution

mcdiag

Use a diagonal covariance matrix

mcnoscale

Do not scale the algorithm during

mcu

Use a uniform distribution as proposal distribution

hybrid

Use the hybrid method

hynstep N

Mean number of steps for the leapfrog method

hyeps X

The stepsize for the leapfrog method [X numeric and > 0]

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.