Model parameters class

Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.

SCAPars(object, ...)

# S4 method for class 'missing'
SCAPars(object, ...)

stkmodel(object, ...)

# S4 method for class 'SCAPars'
stkmodel(object)

n1model(object, ...)

# S4 method for class 'SCAPars'
n1model(object)

srmodel(object, ...)

# S4 method for class 'SCAPars'
srmodel(object)

fmodel(object, ...)

# S4 method for class 'SCAPars'
fmodel(object)

qmodel(object, ...)

# S4 method for class 'SCAPars'
qmodel(object)

qMod(object, ...)

# S4 method for class 'SCAPars'
qMod(object)

vmodel(object, ...)

# S4 method for class 'SCAPars'
vmodel(object)

vMod(object, ...)

# S4 method for class 'SCAPars'
vMod(object)

srPars(object, ...)

# S4 method for class 'SCAPars'
srPars(object)

srCovar(object, ...)

# S4 method for class 'SCAPars'
srCovar(object)

srFrml(object, ...)

# S4 method for class 'SCAPars'
srFrml(object)

fPars(object, ...)

# S4 method for class 'SCAPars'
fPars(object)

fCovar(object, ...)

# S4 method for class 'SCAPars'
fCovar(object)

fFrml(object, ...)

# S4 method for class 'SCAPars'
fFrml(object)

qPars(object, ...)

# S4 method for class 'SCAPars'
qPars(object)

qCovar(object, ...)

# S4 method for class 'SCAPars'
qCovar(object)

qFrml(object, ...)

# S4 method for class 'SCAPars'
qFrml(object)

vPars(object, ...)

# S4 method for class 'SCAPars'
vPars(object)

vCovar(object, ...)

# S4 method for class 'SCAPars'
vCovar(object)

vFrml(object, ...)

# S4 method for class 'SCAPars'
vFrml(object)

# S4 method for class 'SCAPars'
m(object)

# S4 method for class 'SCAPars'
wt(object)

# S4 method for class 'SCAPars'
propagate(object, iter, fill.iter = TRUE)

# S4 method for class 'SCAPars'
iter(obj, it)

Arguments

object

object of relevant class (see signature of method)

...

additional argument list that might never be used

iter

the number of iterations to create

fill.iter

should the new iterations be filled with values (TRUE) or NAs (FALSE)

obj

the object to be subset

it

iteration to be extracted

Slot

stkmodel

parameters related to stock dynamics

qmodel

paramaters related to catchability of tunning fleets

vmodel

paramaters related to the variance model

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.