S4 class a4aFitMCMC

The a4aFitMCMC class extends a4aFitSA to store information about the MCMC run.

a4aFitMCMC(...)

a4aFitMCMC(...)

# S4 method for class 'a4aFitMCMC'
a4aFitSA(object, ...)

# S4 method for class 'a4aFitMCMC'
a4aFit(object, ...)

as.mcmc(x, ...)

# S4 method for class 'a4aFitMCMC'
as.mcmc(x, ...)

burnin(object, ...)

# S4 method for class 'a4aFitMCMC'
burnin(object, burnin)

a4aFitMCMCs(object, ...)

# S4 method for class 'list'
a4aFitMCMCs(object, ...)

# S4 method for class 'a4aFitMCMC'
a4aFitMCMCs(object, ...)

# S4 method for class 'missing'
a4aFitMCMCs(object, ...)

Arguments

...

additional argument list that might never be used

object

object of relevant class (see signature of method)

x

an object to be coerced into mcmc

burnin

a numeric with the number of iterations to be removed

Slots

name

A character vector for the object name.

desc

A textual description of the object contents.

range

A named numeric vector with various values of quant and year ranges, plusgroup, fishing mortality ranges, etc.

call

The function call

clock

Information on call duration

fitSumm

Fit summary

stock.n

Estimates of stock numbers-at-age

harvest

Estimates of fishing mortality at age

catch.n

Estimates of catch numbers-at-age

index

Estimates of survey or CPUE indices-at-age

mcmc

An object of class SCAMCMC with information about the MCMC run

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.

Examples

data(ple4)
data(ple4.index)

obj <- sca(stock=ple4, indices=FLIndices(ple4.index), fit="assessment")
#> Warning: *** ~te(age, year, k = c(6, 30), bs = "tp") + s(age, k = 6) has 4 too many parameter(s)!!
#>     i will remove the redundant ones:
#> 	s(age).2
#> 	s(age).3
#> 	s(age).4
#> 	s(age).5
obj
#> a4a model fit for: PLE 
#> 
#> Call:
#> .local(stock = stock, indices = indices, fit = "assessment")
#> 
#> Time used:
#>  Pre-processing     Running a4a Post-processing           Total 
#>      0.24659991      3.85717654      0.04121041      4.14498687 
#> 
#> Submodels:
#> 	 fmodel: ~te(age, year, k = c(6, 30), bs = "tp") + s(age, k = 6)
#> 	srmodel: ~factor(year)
#> 	n1model: ~s(age, k = 3)
#> 	 qmodel:
#> 	   BTS-Combined (all): ~s(age, k = 6)
#> 	 vmodel:
#> 	   catch:              ~s(age, k = 3)
#> 	   BTS-Combined (all): ~1

slotNames(obj)
#>  [1] "pars"    "call"    "clock"   "fitSumm" "stock.n" "harvest" "catch.n"
#>  [8] "index"   "name"    "desc"    "range"  
clock(obj)
#>  Pre-processing     Running a4a Post-processing           Total 
#>      0.24659991      3.85717654      0.04121041      4.14498687 
fitSumm(obj)
#>              iters
#>                           1
#>   nopar        2.540000e+02
#>   nlogl       -1.015242e+03
#>   maxgrad      1.198324e-04
#>   nobs         8.300000e+02
#>   gcv          1.072679e-01
#>   convergence  0.000000e+00
#>   accrate                NA
#>   nlogl_comp1 -1.070550e+03
#>   nlogl_comp2  5.531180e+01

flq <- stock.n(obj)
is(flq)
#> [1] "FLQuant"   "FLArray"   "array"     "structure" "vector"   
flq <- index(obj)
is(flq)
#> [1] "FLQuants" "FLlst"    "list"     "vector"  

logLik(obj)
#> 'log Lik.' 1015.242 (df=254)
AIC(obj)
#> [1] -1522.484
BIC(obj)
#> [1] -323.2414

is(pars(obj))
#> [1] "SCAPars"