Stock parameters class

Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.

a4aStkParams(object, ...)

# S4 method for missing
a4aStkParams(object, ...)

# S4 method for a4aStkParams
m(object)

# S4 method for a4aStkParams
wt(object)

# S4 method for a4aStkParams
mat(object)

fMod(object, ...)

# S4 method for a4aStkParams
fMod(object)

fMod(object) <- value

# S4 method for a4aStkParams,formula
fMod(object) <- value

n1Mod(object, ...)

# S4 method for a4aStkParams
n1Mod(object)

n1Mod(object) <- value

# S4 method for a4aStkParams,formula
n1Mod(object) <- value

srMod(object, ...)

# S4 method for a4aStkParams
srMod(object)

srMod(object) <- value

# S4 method for a4aStkParams,formula
srMod(object) <- value

# S4 method for a4aStkParams
params(object)

# S4 method for a4aStkParams,FLPar
params(object) <- value

coefficients(object, ...)

# S4 method for a4aStkParams
coefficients(object)

coefficients(object) <- value

# S4 method for a4aStkParams,FLPar
coefficients(object) <- value

# S4 method for a4aStkParams
distr(object)

# S4 method for a4aStkParams,character
distr(object) <- value

# S4 method for a4aStkParams
vcov(object)

# S4 method for a4aStkParams,array
vcov(object) <- value

# S4 method for a4aStkParams
propagate(object, iter, fill.iter = TRUE)

# S4 method for a4aStkParams
iter(obj, it)

Arguments

object

object of relevant class (see signature of method)

...

additional argument list that might never be used

value

the new object

iter

the number of iterations to create

fill.iter

should the new iterations be filled with values (TRUE) or NAs (FALSE)

obj

the object to be subset

it

iteration to be extracted

Slot

fMod

F submodel formula

n1Mod

first year N formula

srMod

stock-recruitment submodel formula

params

FLPar with parameters

vcov

array with variance-covariance

centering

centering values numeric

distr

statistical distribution character

m

natural mortality FLQuant

units

data units character

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.