Methods to generate FLStock objects

This method computes the FLStock slots consistently with the information provided by the FLQuant. It requires two of the triplet R/C/F to compute the third consistent with Baranov and survival's equations.

genFLStock(object, R, C, F, ...)

# S4 method for FLStock,FLQuant,FLQuant,missing
genFLStock(object, R, C, F,
  ...)

# S4 method for FLStock,missing,FLQuant,FLQuant
genFLStock(object, R, C, F,
  ...)

# S4 method for FLStock,FLQuant,missing,FLQuant
genFLStock(object, R, C, F,
  ...)

Arguments

object
an FLStock

R
an FLQuant with iterations or missing

C
an FLQuant with iterations or missing

F
an FLQuant with iterations or missing

...
additional argument list that might not ever be used.

Value

an FLStock