Run several stock assessments in a single run

Internal method to run several stock assessment fits with different stocks, indices and submodels

scas(
  stocks,
  indicess,
  fmodel = missing,
  qmodel = missing,
  srmodel = missing,
  n1model = missing,
  vmodel = missing,
  combination.all = FALSE,
  workers = 1,
  ...
)

Arguments

stocks

an FLStocks object containing catch and stock information

indicess

a list of FLIndices objects containing survey indices

fmodel

a list of fmodel objects, each with a formula object depicting the model for log fishing mortality at age

qmodel

a list of qmodel objects, each with a list of formula objects depicting the models for log survey catchability at age

srmodel

a list of srmodel objects, each with a formula object depicting the model for log recruitment

n1model

a list of n1model objects, each with a formula object depicting the model for the first year of catch data

vmodel

a list of vmodel objects, each with a list of formula objects depicting the models for log survey and log fishing mortality variance

combination.all

bolean parameter (default is FALSE) to define if a full factorial across all stocks, indices, and submodel is run or just a sequence of runs.

...

all other arguments to be passed to sca

stock

an FLStocks object, each component with a FLStock object containing catch and stock information

Value

an a4aFits or a4aFitSAs or a4aFitMCMCs depending on the argument fit