Model parameters class

Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.

SCAPars(object, ...)

# S4 method for missing
SCAPars(object, ...)

stkmodel(object, ...)

# S4 method for SCAPars
stkmodel(object)

n1model(object, ...)

# S4 method for SCAPars
n1model(object)

srmodel(object, ...)

# S4 method for SCAPars
srmodel(object)

fmodel(object, ...)

# S4 method for SCAPars
fmodel(object)

qmodel(object, ...)

# S4 method for SCAPars
qmodel(object)

vmodel(object, ...)

# S4 method for SCAPars
vmodel(object)

srPars(object, ...)

# S4 method for SCAPars
srPars(object)

srCovar(object, ...)

# S4 method for SCAPars
srCovar(object)

srFrml(object, ...)

# S4 method for SCAPars
srFrml(object)

fPars(object, ...)

# S4 method for SCAPars
fPars(object)

fCovar(object, ...)

# S4 method for SCAPars
fCovar(object)

fFrml(object, ...)

# S4 method for SCAPars
fFrml(object)

qPars(object, ...)

# S4 method for SCAPars
qPars(object)

qCovar(object, ...)

# S4 method for SCAPars
qCovar(object)

qFrml(object, ...)

# S4 method for SCAPars
qFrml(object)

vPars(object, ...)

# S4 method for SCAPars
vPars(object)

vCovar(object, ...)

# S4 method for SCAPars
vCovar(object)

vFrml(object, ...)

# S4 method for SCAPars
vFrml(object)

# S4 method for SCAPars
m(object)

# S4 method for SCAPars
wt(object)

Arguments

object

object of relevant class (see signature of method)

...

additional argument list that might never be used

Slot

stkmodel

parameters related to stock dynamics

qmodel

paramaters related to catchability of tunning fleets

vmodel

paramaters related to the variance model

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.