S4 class

The a4aFit class was built to store the a4a stock assessment fits.

a4aFit(object, ...)

# S4 method for missing
a4aFit(object, ...)

clock(object, ...)

# S4 method for a4aFit
clock(object)

fitSumm(object, ...)

# S4 method for a4aFit
fitSumm(object)

# S4 method for a4aFit
stock.n(object)

# S4 method for a4aFit,ANY
harvest(object)

# S4 method for a4aFit
catch.n(object)

# S4 method for a4aFit
index(object)

# S4 method for a4aFit
show(object)

# S4 method for a4aFit
logLik(object, ...)

Arguments

object

object of relevant class (see signature of method)

...

additional argument list that might never be used

Slots

call

The function call

clock

Information on call duration

fitSumm

Fit summary

stock.n

Estimates of stock numbers-at-age

harvest

Estimates of fishing mortality at age

catch.n

Estimates of catch numbers-at-age

index

Estimates of survey or CPUE indices-at-age

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.

Examples

data(ple4)
data(ple4.index)

obj <- sca(stock=ple4, indices=FLIndices(ple4.index))
obj

#> a4a model fit for: Plaice in IV 
#> 
#> Call:
#> .local(stock = stock, indices = indices, fmodel = ..1, qmodel = ..2, 
#>     srmodel = ..3, n1model = ..4, vmodel = ..5, fit = ..6)
#> 
#> Time used:
#>  Pre-processing     Running a4a Post-processing           Total 
#>       0.9497073       1.7033591       0.1171367       2.7702031 

slotNames(obj)

#>  [1] "call"    "clock"   "fitSumm" "stock.n" "harvest" "catch.n" "index"  
#>  [8] "name"    "desc"    "range"  

clock(obj)

#>  Pre-processing     Running a4a Post-processing           Total 
#>       0.9497073       1.7033591       0.1171367       2.7702031 

fitSumm(obj)

#>              iters
#>                           1
#>   nopar        1.780000e+02
#>   nlogl       -7.771360e+01
#>   maxgrad      5.430380e-06
#>   nobs         7.100000e+02
#>   gcv          4.824039e-02
#>   convergence            NA
#>   accrate                NA

flq <- stock.n(obj)
is(flq)

#> [1] "FLQuant"   "FLArray"   "array"     "structure" "vector"   

flq <- index(obj)
is(flq)

#> [1] "FLQuants" "FLlst"    "list"     "vector"  

logLik(obj)

#> 'log Lik.' 77.7136 (df=178)

AIC(obj)

#> [1] 200.5728

BIC(obj)

#> [1] 1013.19