Natural mortality class

Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.

a4aM(object, ...)

# S4 method for missing
a4aM(object, ...)

# S4 method for a4aM
show(object)

shape(object, ...)

# S4 method for a4aM
shape(object)

shape(object) <- value

# S4 method for a4aM
shape(object) <- value

level(object, ...)

# S4 method for a4aM
level(object)

level(object) <- value

# S4 method for a4aM
level(object) <- value

trend(object, ...)

# S4 method for a4aM
trend(object)

trend(object) <- value

# S4 method for a4aM
trend(object) <- value

Arguments

object
object of relevant class (see signature of method)

...
additional argument list that might never be used

value
the new object

Slot

shape

the shape of M by age

level

the mean level of M over a range of ages, which will be used to scale the shape

trend

the yearly trend in M

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing, but not for populating any slot.

Examples

mod1 <- FLModelSim(model=~exp(-age-0.5))

#> Error in validObject(.Object): invalid class “FLModelSim” object: Object is not valid. Check that the names of the parameters in the params matrix and the vcov match the names of the formula parameters.

mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)

#> Error in a4aM(shape = mod1, level = mod2): object 'mod1' not found