+ methods

Update FLStock and FLIndex objects with stock assessment results.

# S4 method for FLStock,a4aFit
+(e1, e2)

# S4 method for FLIndices,a4aFit
+(e1, e2)

Arguments

e1
the original FLStock or FLIndex object

e2
a a4aFit object from where the new FLStock or FLIndex slots will be extracted.

Details

If both objects have the same number of iterations, the FLStock slots will be replaced by the a4aFit slots, in the case of 1 iter, or a4aFitSA slots, in the case of n iters. If one of the objects has 1 iter and the other n, the method will simulate using the fit results from the a4aFitSA object to update the slots of the FLStock object.