Assorted methods needed by FLa4a

Assorted methods needed by FLa4a

getYidx(object, ...)

# S4 method for FLQuant
getYidx(object, year)

is.empty(object)

niters(object, ...)

# S4 method for FLModelSim
niters(object)

# S4 method for a4aGr
niters(object)

# S4 method for a4aStkParams
dims(obj)

Arguments

object

object of relevant class (see signature of method)

...

additional argument list that might never be used

year

numeric with year to be extracted

obj

an object

getYidx

Gets an FLQuant's numeric id for a vector of "years". For internal use and not very interesting for users. It takes an FLQuant object and vector of years and returns a numeric vector that can be used to subset the FLQuant.

is.empty

Method is.empty checks if an object is empty. It takes any object and returns a logical, TRUE, if the object is of length 0.

niters

Compute number of iterations. Takes an object of any FLR class and returns a numeric.

dims

Extracts the dims of the parameters.

Examples

#Example use of getYidx:
data(ple4)
flq <- catch(ple4)
getYidx(flq, 2000:2004)

#> [1] 44 45 46 47 48

flq[, getYidx(flq, 2000:2004)]

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   2000   2001   2002   2003   2004  
#>   all 126142 182578 125884 145390 117702
#> 
#> units:  t 

#Example use of is.empty:
is.empty(list())

#> [1] TRUE

is.empty(list(a=2))

#> [1] FALSE

#Example use of niters:
mm <- matrix(NA, ncol=3, nrow=3)
diag(mm) <- c(50, 0.001,0.001)
mm[upper.tri(mm)] <- mm[lower.tri(mm)] <- c(0.1,0.01,0.00004)
md <- ~linf*(1-exp(-k*(t-t0)))
imd <- ~t0-1/k*log(1-len/linf)
prs <- FLPar(linf=58.5, k=0.086, t0=0.001, units=c("cm","yr^-1","yr"))
vbObj <- a4aGr(grMod=md, grInvMod=imd, params=prs, vcov=mm, distr="norm")
# Generate 100 sample sets
vbObj <- mvrnorm(100,vbObj)
niters(vbObj)

#> [1] 100

#Example use of dims:
dims(FLPar())

#> $iter
#> [1] 1
#> 
#> $params
#> [1] ""
#>