natural mortality

Method to compute natural mortality.

# S4 method for a4aM
m(object, grMod = "missing", ...)

Arguments

object
a a4aM object

grMod
a a4aGr object from which the growth parameter K can be extracted

...
placeholder for covariates of the models. The names must match formula variables (not parameters), with the exception of the a4aGr individual growth model. To use a growth model, it must be called grMod and be of class a4aGr, in which case the parameters will be matched. The main objective is to be able to use K from von Bertalanffy models in M.

Value

an FLQuant object

Details

The method uses the range slot to define the quant and year dimensions of the resulting M FLQuant. The name for the quant dimension is taken as the name of a variable that is present in the shape formula, but not in the params slot of the shape model. If more than one such variable exists, then there is a problem with the shape model definition.

Examples

age <- 0:15
k <- 0.4
shp <- eval(as.list(~exp(-age-0.5))[[2]], envir=list(age=age))
lvl <- eval(as.list(~1.5*k)[[2]], envir=list(k=k))
M <- shp*lvl/mean(shp)
# Now set up an equivalent a4aM object
mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)
  # set up the age range for the object...
  range(m1, c("min", "max")) <- c(0,15)
  # ...and the age range for mbar
  range(m1, c("minmbar", "maxmbar")) <- c(0,15)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   0  6.0684e+00
#>   1  2.2324e+00
#>   2  8.2126e-01
#>   3  3.0213e-01
#>   4  1.1115e-01
#>   5  4.0888e-02
#>   6  1.5042e-02
#>   7  5.5336e-03
#>   8  2.0357e-03
#>   9  7.4889e-04
#>   10 2.7550e-04
#>   11 1.0135e-04
#>   12 3.7285e-05
#>   13 1.3716e-05
#>   14 5.0460e-06
#>   15 1.8563e-06
#> 
#> units:  NA 

mean(m(m1)[ac(0:15)])

#> [1] 0.6

all.equal(M, c(m(m1)))

#> [1] TRUE

# another example m
range(m1, c("min", "max")) <- c(2,15)
range(m1, c("minmbar", "maxmbar")) <- c(2,4)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   2  1.1974e+00
#>   3  4.4051e-01
#>   4  1.6206e-01
#>   5  5.9617e-02
#>   6  2.1932e-02
#>   7  8.0682e-03
#>   8  2.9681e-03
#>   9  1.0919e-03
#>   10 4.0169e-04
#>   11 1.4778e-04
#>   12 5.4363e-05
#>   13 1.9999e-05
#>   14 7.3573e-06
#>   15 2.7066e-06
#> 
#> units:  NA 

mean(m(m1)[ac(2:4)])

#> [1] 0.6

# example with specified iters (i.e. not simulated from a statistical distribution)...
mod2 <- FLModelSim(model=~k^0.66*t^0.57,
  params=FLPar(matrix(c(0.4,10,0.5,11), ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
  vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.002), dim=c(2,2,2)))
m2 <- a4aM(shape=mod1, level=mod2)
range(m2, c("min", "max")) <- c(2,10)
m(m2)

#> An object of class "FLQuant"
#> iters:  2 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                  
#>   2  2.2560020(0.336016)
#>   3  0.8299367(0.123613)
#>   4  0.3053167(0.045475)
#>   5  0.1123197(0.016729)
#>   6  0.0413201(0.006154)
#>   7  0.0152008(0.002264)
#>   8  0.0055921(0.000833)
#>   9  0.0020572(0.000306)
#>   10 0.0007568(0.000113)
#> 
#> units:  NA 

# ...and with randomly generated iters (based on the medians for params(mod2) and vcov(mod2))
m3 <- a4aM(shape=mod1, level=mvrnorm(100, mod2))
range(m3, c("min", "max")) <- c(0,15)
m(m3)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                   
#>   0  2.2955e+00(2.13e-01)
#>   1  8.4446e-01(7.84e-02)
#>   2  3.1066e-01(2.88e-02)
#>   3  1.1429e-01(1.06e-02)
#>   4  4.2043e-02(3.90e-03)
#>   5  1.5467e-02(1.44e-03)
#>   6  5.6899e-03(5.28e-04)
#>   7  2.0932e-03(1.94e-04)
#>   8  7.7005e-04(7.15e-05)
#>   9  2.8328e-04(2.63e-05)
#>   10 1.0421e-04(9.67e-06)
#>   11 3.8338e-05(3.56e-06)
#>   12 1.4104e-05(1.31e-06)
#>   13 5.1885e-06(4.82e-07)
#>   14 1.9088e-06(1.77e-07)
#>   15 7.0219e-07(6.52e-08)
#> 
#> units:  NA 

# example with a trend
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mvrnorm(100, mod2), trend=mod3,
  range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))
m(mObj, v=1:4)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000                 2001                 2002                
#>   0  2.3406e+00(2.74e-01) 2.4521e+00(2.88e-01) 2.5635e+00(3.01e-01)
#>   1  8.6106e-01(1.01e-01) 9.0206e-01(1.06e-01) 9.4306e-01(1.11e-01)
#>   2  3.1677e-01(3.71e-02) 3.3185e-01(3.89e-02) 3.4693e-01(4.07e-02)
#>   3  1.1653e-01(1.37e-02) 1.2208e-01(1.43e-02) 1.2763e-01(1.50e-02)
#>   4  4.2870e-02(5.03e-03) 4.4911e-02(5.27e-03) 4.6952e-02(5.51e-03)
#>   5  1.5771e-02(1.85e-03) 1.6522e-02(1.94e-03) 1.7273e-02(2.03e-03)
#>   6  5.8018e-03(6.80e-04) 6.0780e-03(7.13e-04) 6.3543e-03(7.45e-04)
#>   7  2.1343e-03(2.50e-04) 2.2360e-03(2.62e-04) 2.3376e-03(2.74e-04)
#>   8  7.8518e-04(9.21e-05) 8.2257e-04(9.65e-05) 8.5996e-04(1.01e-04)
#>   9  2.8885e-04(3.39e-05) 3.0261e-04(3.55e-05) 3.1636e-04(3.71e-05)
#>   10 1.0626e-04(1.25e-05) 1.1132e-04(1.31e-05) 1.1638e-04(1.36e-05)
#>   11 3.9092e-05(4.58e-06) 4.0953e-05(4.80e-06) 4.2815e-05(5.02e-06)
#>   12 1.4381e-05(1.69e-06) 1.5066e-05(1.77e-06) 1.5751e-05(1.85e-06)
#>   13 5.2905e-06(6.20e-07) 5.5424e-06(6.50e-07) 5.7944e-06(6.79e-07)
#>   14 1.9463e-06(2.28e-07) 2.0390e-06(2.39e-07) 2.1316e-06(2.50e-07)
#>   15 7.1599e-07(8.40e-08) 7.5009e-07(8.80e-08) 7.8418e-07(9.20e-08)
#>     year
#> age  2003                
#>   0  2.6750e+00(3.14e-01)
#>   1  9.8407e-01(1.15e-01)
#>   2  3.6202e-01(4.24e-02)
#>   3  1.3318e-01(1.56e-02)
#>   4  4.8994e-02(5.74e-03)
#>   5  1.8024e-02(2.11e-03)
#>   6  6.6306e-03(7.77e-04)
#>   7  2.4393e-03(2.86e-04)
#>   8  8.9735e-04(1.05e-04)
#>   9  3.3012e-04(3.87e-05)
#>   10 1.2144e-04(1.42e-05)
#>   11 4.4677e-05(5.24e-06)
#>   12 1.6436e-05(1.93e-06)
#>   13 6.0463e-06(7.09e-07)
#>   14 2.2243e-06(2.61e-07)
#>   15 8.1828e-07(9.59e-08)
#> 
#> units:  NA