natural mortality

Method to compute natural mortality.

# S4 method for a4aM
m(object, grMod = "missing", ...)

Arguments

object

a a4aM object

grMod

a a4aGr object from which the growth parameter K can be extracted

...

placeholder for covariates of the models. The names must match formula variables (not parameters), with the exception of the a4aGr individual growth model. To use a growth model, it must be called grMod and be of class a4aGr, in which case the parameters will be matched. The main objective is to be able to use K from von Bertalanffy models in M.

Value

an FLQuant object

Details

The method uses the range slot to define the quant and year dimensions of the resulting M FLQuant. The name for the quant dimension is taken as the name of a variable that is present in the shape formula, but not in the params slot of the shape model. If more than one such variable exists, then there is a problem with the shape model definition.

Examples

age <- 0:15
k <- 0.4
shp <- eval(as.list(~exp(-age-0.5))[[2]], envir=list(age=age))
lvl <- eval(as.list(~1.5*k)[[2]], envir=list(k=k))
M <- shp*lvl/mean(shp)
# Now set up an equivalent a4aM object
mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)
  # set up the age range for the object...
  rngquant(m1) <- c(0,15)
  # ...and the age range for mbar
  rngmbar(m1)<-c(0,15)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   0  6.0684e+00
#>   1  2.2324e+00
#>   2  8.2126e-01
#>   3  3.0213e-01
#>   4  1.1115e-01
#>   5  4.0888e-02
#>   6  1.5042e-02
#>   7  5.5336e-03
#>   8  2.0357e-03
#>   9  7.4889e-04
#>   10 2.7550e-04
#>   11 1.0135e-04
#>   12 3.7285e-05
#>   13 1.3716e-05
#>   14 5.0460e-06
#>   15 1.8563e-06
#> 
#> units:  NA 

mean(m(m1)[ac(0:15)])

#> [1] 0.6

all.equal(M, c(m(m1)))

#> [1] TRUE

# another example m
rngquant(m1) <- c(2,10)
rngmbar(m1) <- c(2,4)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   2  1.19743372
#>   3  0.44051125
#>   4  0.16205503
#>   5  0.05961671
#>   6  0.02193176
#>   7  0.00806824
#>   8  0.00296814
#>   9  0.00109192
#>   10 0.00040169
#> 
#> units:  NA 

mean(m(m1)[ac(2:4)])

#> [1] 0.6

# example with specified iters (i.e. not simulated from a statistical distribution)...
mod2 <- FLModelSim(model=~k^0.66*t^0.57,
  params=FLPar(matrix(c(0.4,10,0.5,11), ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
  vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.002), dim=c(2,2,2)))
m2 <- a4aM(shape=mod1, level=mod2)
rngquant(m2) <- c(0,10)
m(m2)

#> An object of class "FLQuant"
#> iters:  2 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                   
#>   0  2.25600198(3.36e-01)
#>   1  0.82993675(1.24e-01)
#>   2  0.30531667(4.55e-02)
#>   3  0.11231972(1.67e-02)
#>   4  0.04132012(6.15e-03)
#>   5  0.01520082(2.26e-03)
#>   6  0.00559207(8.33e-04)
#>   7  0.00205721(3.06e-04)
#>   8  0.00075680(1.13e-04)
#>   9  0.00027841(4.15e-05)
#>   10 0.00010242(1.53e-05)
#> 
#> units:  NA 

# ...and with randomly generated iters (based on the medians for params(mod2) and vcov(mod2))
m3 <- a4aM(shape=mod1, level=mvrnorm(100, mod2))
rngquant(m3) <- c(0,15)
m(m3)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                   
#>   0  2.2880e+00(2.49e-01)
#>   1  8.4171e-01(9.17e-02)
#>   2  3.0965e-01(3.37e-02)
#>   3  1.1391e-01(1.24e-02)
#>   4  4.1906e-02(4.57e-03)
#>   5  1.5417e-02(1.68e-03)
#>   6  5.6714e-03(6.18e-04)
#>   7  2.0864e-03(2.27e-04)
#>   8  7.6754e-04(8.36e-05)
#>   9  2.8236e-04(3.08e-05)
#>   10 1.0388e-04(1.13e-05)
#>   11 3.8214e-05(4.16e-06)
#>   12 1.4058e-05(1.53e-06)
#>   13 5.1717e-06(5.63e-07)
#>   14 1.9025e-06(2.07e-07)
#>   15 6.9991e-07(7.63e-08)
#> 
#> units:  NA 

# example with a trend
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mvrnorm(100, mod2), trend=mod3,
  range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))
m(mObj, v=1:4)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000                 2001                 2002                
#>   0  2.3565e+00(2.39e-01) 2.4687e+00(2.50e-01) 2.5810e+00(2.61e-01)
#>   1  8.6692e-01(8.78e-02) 9.0820e-01(9.19e-02) 9.4948e-01(9.61e-02)
#>   2  3.1892e-01(3.23e-02) 3.3411e-01(3.38e-02) 3.4929e-01(3.54e-02)
#>   3  1.1732e-01(1.19e-02) 1.2291e-01(1.24e-02) 1.2850e-01(1.30e-02)
#>   4  4.3161e-02(4.37e-03) 4.5217e-02(4.58e-03) 4.7272e-02(4.79e-03)
#>   5  1.5878e-02(1.61e-03) 1.6634e-02(1.68e-03) 1.7390e-02(1.76e-03)
#>   6  5.8412e-03(5.91e-04) 6.1194e-03(6.19e-04) 6.3976e-03(6.48e-04)
#>   7  2.1489e-03(2.18e-04) 2.2512e-03(2.28e-04) 2.3535e-03(2.38e-04)
#>   8  7.9053e-04(8.00e-05) 8.2817e-04(8.38e-05) 8.6582e-04(8.77e-05)
#>   9  2.9082e-04(2.94e-05) 3.0467e-04(3.08e-05) 3.1852e-04(3.22e-05)
#>   10 1.0699e-04(1.08e-05) 1.1208e-04(1.13e-05) 1.1718e-04(1.19e-05)
#>   11 3.9358e-05(3.98e-06) 4.1232e-05(4.17e-06) 4.3106e-05(4.36e-06)
#>   12 1.4479e-05(1.47e-06) 1.5168e-05(1.54e-06) 1.5858e-05(1.61e-06)
#>   13 5.3265e-06(5.39e-07) 5.5802e-06(5.65e-07) 5.8338e-06(5.91e-07)
#>   14 1.9595e-06(1.98e-07) 2.0528e-06(2.08e-07) 2.1461e-06(2.17e-07)
#>   15 7.2087e-07(7.30e-08) 7.5519e-07(7.65e-08) 7.8952e-07(7.99e-08)
#>     year
#> age  2003                
#>   0  2.6932e+00(2.73e-01)
#>   1  9.9076e-01(1.00e-01)
#>   2  3.6448e-01(3.69e-02)
#>   3  1.3409e-01(1.36e-02)
#>   4  4.9327e-02(4.99e-03)
#>   5  1.8146e-02(1.84e-03)
#>   6  6.6757e-03(6.76e-04)
#>   7  2.4559e-03(2.49e-04)
#>   8  9.0346e-04(9.15e-05)
#>   9  3.3236e-04(3.36e-05)
#>   10 1.2227e-04(1.24e-05)
#>   11 4.4981e-05(4.55e-06)
#>   12 1.6547e-05(1.68e-06)
#>   13 6.0875e-06(6.16e-07)
#>   14 2.2395e-06(2.27e-07)
#>   15 8.2385e-07(8.34e-08)
#> 
#> units:  NA