natural mortality

Method to compute natural mortality.

# S4 method for a4aM
m(object, grMod = "missing", ...)

Arguments

object

a a4aM object

grMod

a a4aGr object from which the growth parameter K can be extracted

...

placeholder for covariates of the models. The names must match formula variables (not parameters), with the exception of the a4aGr individual growth model. To use a growth model, it must be called grMod and be of class a4aGr, in which case the parameters will be matched. The main objective is to be able to use K from von Bertalanffy models in M.

Value

an FLQuant object

Details

The method uses the range slot to define the quant and year dimensions of the resulting M FLQuant. The name for the quant dimension is taken as the name of a variable that is present in the shape formula, but not in the params slot of the shape model. If more than one such variable exists, then there is a problem with the shape model definition.

Examples

age <- 0:15
k <- 0.4
shp <- eval(as.list(~exp(-age-0.5))[[2]], envir=list(age=age))
lvl <- eval(as.list(~1.5*k)[[2]], envir=list(k=k))
M <- shp*lvl/mean(shp)
# Now set up an equivalent a4aM object
mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)
  # set up the age range for the object...
  range(m1, c("min", "max")) <- c(0,15)
  # ...and the age range for mbar
  range(m1, c("minmbar", "maxmbar")) <- c(0,15)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   0  6.0684e+00
#>   1  2.2324e+00
#>   2  8.2126e-01
#>   3  3.0213e-01
#>   4  1.1115e-01
#>   5  4.0888e-02
#>   6  1.5042e-02
#>   7  5.5336e-03
#>   8  2.0357e-03
#>   9  7.4889e-04
#>   10 2.7550e-04
#>   11 1.0135e-04
#>   12 3.7285e-05
#>   13 1.3716e-05
#>   14 5.0460e-06
#>   15 1.8563e-06
#> 
#> units:  NA 

mean(m(m1)[ac(0:15)])

#> [1] 0.6

all.equal(M, c(m(m1)))

#> [1] TRUE

# another example m
range(m1, c("min", "max")) <- c(2,15)
range(m1, c("minmbar", "maxmbar")) <- c(2,4)
m(m1)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0         
#>   2  1.1974e+00
#>   3  4.4051e-01
#>   4  1.6206e-01
#>   5  5.9617e-02
#>   6  2.1932e-02
#>   7  8.0682e-03
#>   8  2.9681e-03
#>   9  1.0919e-03
#>   10 4.0169e-04
#>   11 1.4778e-04
#>   12 5.4363e-05
#>   13 1.9999e-05
#>   14 7.3573e-06
#>   15 2.7066e-06
#> 
#> units:  NA 

mean(m(m1)[ac(2:4)])

#> [1] 0.6

# example with specified iters (i.e. not simulated from a statistical distribution)...
mod2 <- FLModelSim(model=~k^0.66*t^0.57,
  params=FLPar(matrix(c(0.4,10,0.5,11), ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
  vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.002), dim=c(2,2,2)))
m2 <- a4aM(shape=mod1, level=mod2)
range(m2, c("min", "max")) <- c(2,10)
m(m2)

#> An object of class "FLQuant"
#> iters:  2 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                  
#>   2  2.2560020(0.336016)
#>   3  0.8299367(0.123613)
#>   4  0.3053167(0.045475)
#>   5  0.1123197(0.016729)
#>   6  0.0413201(0.006154)
#>   7  0.0152008(0.002264)
#>   8  0.0055921(0.000833)
#>   9  0.0020572(0.000306)
#>   10 0.0007568(0.000113)
#> 
#> units:  NA 

# ...and with randomly generated iters (based on the medians for params(mod2) and vcov(mod2))
m3 <- a4aM(shape=mod1, level=mvrnorm(100, mod2))
range(m3, c("min", "max")) <- c(0,15)
m(m3)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  0                   
#>   0  2.2769e+00(1.98e-01)
#>   1  8.3761e-01(7.27e-02)
#>   2  3.0814e-01(2.68e-02)
#>   3  1.1336e-01(9.85e-03)
#>   4  4.1702e-02(3.62e-03)
#>   5  1.5341e-02(1.33e-03)
#>   6  5.6438e-03(4.90e-04)
#>   7  2.0762e-03(1.80e-04)
#>   8  7.6380e-04(6.63e-05)
#>   9  2.8099e-04(2.44e-05)
#>   10 1.0337e-04(8.98e-06)
#>   11 3.8028e-05(3.30e-06)
#>   12 1.3990e-05(1.21e-06)
#>   13 5.1465e-06(4.47e-07)
#>   14 1.8933e-06(1.64e-07)
#>   15 6.9650e-07(6.05e-08)
#> 
#> units:  NA 

# example with a trend
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mvrnorm(100, mod2), trend=mod3,
  range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))
m(mObj, v=1:4)

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000                 2001                 2002                
#>   0  2.3681e+00(2.58e-01) 2.4809e+00(2.71e-01) 2.5937e+00(2.83e-01)
#>   1  8.7119e-01(9.50e-02) 9.1268e-01(9.96e-02) 9.5416e-01(1.04e-01)
#>   2  3.2049e-01(3.50e-02) 3.3576e-01(3.66e-02) 3.5102e-01(3.83e-02)
#>   3  1.1790e-01(1.29e-02) 1.2352e-01(1.35e-02) 1.2913e-01(1.41e-02)
#>   4  4.3374e-02(4.73e-03) 4.5440e-02(4.96e-03) 4.7505e-02(5.18e-03)
#>   5  1.5956e-02(1.74e-03) 1.6716e-02(1.82e-03) 1.7476e-02(1.91e-03)
#>   6  5.8701e-03(6.40e-04) 6.1496e-03(6.71e-04) 6.4291e-03(7.01e-04)
#>   7  2.1595e-03(2.36e-04) 2.2623e-03(2.47e-04) 2.3651e-03(2.58e-04)
#>   8  7.9443e-04(8.67e-05) 8.3226e-04(9.08e-05) 8.7009e-04(9.49e-05)
#>   9  2.9225e-04(3.19e-05) 3.0617e-04(3.34e-05) 3.2009e-04(3.49e-05)
#>   10 1.0751e-04(1.17e-05) 1.1263e-04(1.23e-05) 1.1775e-04(1.28e-05)
#>   11 3.9552e-05(4.31e-06) 4.1436e-05(4.52e-06) 4.3319e-05(4.73e-06)
#>   12 1.4550e-05(1.59e-06) 1.5243e-05(1.66e-06) 1.5936e-05(1.74e-06)
#>   13 5.3528e-06(5.84e-07) 5.6077e-06(6.12e-07) 5.8626e-06(6.39e-07)
#>   14 1.9692e-06(2.15e-07) 2.0630e-06(2.25e-07) 2.1567e-06(2.35e-07)
#>   15 7.2442e-07(7.90e-08) 7.5892e-07(8.28e-08) 7.9341e-07(8.65e-08)
#>     year
#> age  2003                
#>   0  2.7065e+00(2.95e-01)
#>   1  9.9565e-01(1.09e-01)
#>   2  3.6628e-01(4.00e-02)
#>   3  1.3475e-01(1.47e-02)
#>   4  4.9570e-02(5.41e-03)
#>   5  1.8236e-02(1.99e-03)
#>   6  6.7086e-03(7.32e-04)
#>   7  2.4680e-03(2.69e-04)
#>   8  9.0791e-04(9.90e-05)
#>   9  3.3400e-04(3.64e-05)
#>   10 1.2287e-04(1.34e-05)
#>   11 4.5202e-05(4.93e-06)
#>   12 1.6629e-05(1.81e-06)
#>   13 6.1175e-06(6.67e-07)
#>   14 2.2505e-06(2.45e-07)
#>   15 8.2791e-07(9.03e-08)
#> 
#> units:  NA