natural mortality

Method to simulate multivariate normal parameters for an a4aM object.

# S4 method for numeric,a4aM,missing,missing,missing,missing
mvrnorm(n = 1, mu)

Arguments

n
the number of iterations to be generated

mu
an a4aM object

Value

an a4aM object with n iterations

Examples

mod1 <- FLModelSim(model=~exp(-age-0.5))

#> Error in validObject(.Object): invalid class “FLModelSim” object: Object is not valid. Check that the names of the parameters in the params matrix and the vcov match the names of the formula parameters.

mod2 <- FLModelSim(model=~k^0.66*t^0.57, params=FLPar(matrix(c(0.4,10,0.5,11),
 ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
 vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.003), dim=c(2,2,2)))
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mod2, trend=mod3,
 range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))

#> Error in a4aM(shape = mod1, level = mod2, trend = mod3, range = c(min = 0,     max = 15, minyear = 2000, maxyear = 2003, minmbar = 0, maxmbar = 0)): object 'mod1' not found

mObj <- mvrnorm(100, mObj)

#> Error in mvrnorm(100, mObj): object 'mObj' not found

# Generate 100 iterations with no trend over time
  m(mObj, v=c(1,1,1,1))

#> Error in m(mObj, v = c(1, 1, 1, 1)): object 'mObj' not found

# Generate replicates based on iteration-specific multivariate distributions
# (as defined by params() and vcov())
  params(mod2)

#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>             k             t 
#>  0.45(0.0741) 10.50(0.7413) 
#> units:  NA NA 

  vcov(mod2)

#> , , 1
#> 
#>       [,1]  [,2]
#> [1,] 0.004 0.000
#> [2,] 0.000 0.001
#> 
#> , , 2
#> 
#>       [,1]  [,2]
#> [1,] 0.006 0.000
#> [2,] 0.000 0.003
#> 

  m1<-mvrnorm(mod2)
  c(params(m1))

#> [1]  0.4197285 10.0119126  0.4454764 11.0102646

# Generate replicates based on a single multivariate distribution (here the
# median of params() and vcov() is used)
  mvrnorm(2,mod2)

#> An object of class "FLModelSim"
#> Slot "model":
#> ~k^0.66 * t^0.57
#> <environment: 0x192662b8>
#> 
#> Slot "params":
#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>                 k                 t 
#>  0.38612(0.01319) 10.48210(0.00556) 
#> units:  NA NA 
#> 
#> Slot "vcov":
#> numeric(0)
#> 
#> Slot "distr":
#> [1] "norm"
#> 

  m2<-mvrnorm(2,mod2)
  c(params(m2))

#> [1]  0.4670294 10.5645466  0.4312099 10.5337377