natural mortality

Method to simulate multivariate normal parameters for an a4aM object.

# S4 method for numeric,a4aM,missing,missing,missing,missing
mvrnorm(n = 1, mu)

Arguments

n

the number of iterations to be generated

mu

an a4aM object

Value

an a4aM object with n iterations

Examples

mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~k^0.66*t^0.57, params=FLPar(matrix(c(0.4,10,0.5,11),
 ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
 vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.003), dim=c(2,2,2)))
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mod2, trend=mod3,
 range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))
mObj <- mvrnorm(100, mObj)
# Generate 100 iterations with no trend over time
  m(mObj, v=c(1,1,1,1))

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000                 2001                 2002                
#>   0  2.3212e+00(2.01e-01) 2.3212e+00(2.01e-01) 2.3212e+00(2.01e-01)
#>   1  8.5392e-01(7.40e-02) 8.5392e-01(7.40e-02) 8.5392e-01(7.40e-02)
#>   2  3.1414e-01(2.72e-02) 3.1414e-01(2.72e-02) 3.1414e-01(2.72e-02)
#>   3  1.1556e-01(1.00e-02) 1.1556e-01(1.00e-02) 1.1556e-01(1.00e-02)
#>   4  4.2514e-02(3.69e-03) 4.2514e-02(3.69e-03) 4.2514e-02(3.69e-03)
#>   5  1.5640e-02(1.36e-03) 1.5640e-02(1.36e-03) 1.5640e-02(1.36e-03)
#>   6  5.7536e-03(4.99e-04) 5.7536e-03(4.99e-04) 5.7536e-03(4.99e-04)
#>   7  2.1166e-03(1.83e-04) 2.1166e-03(1.83e-04) 2.1166e-03(1.83e-04)
#>   8  7.7867e-04(6.75e-05) 7.7867e-04(6.75e-05) 7.7867e-04(6.75e-05)
#>   9  2.8646e-04(2.48e-05) 2.8646e-04(2.48e-05) 2.8646e-04(2.48e-05)
#>   10 1.0538e-04(9.13e-06) 1.0538e-04(9.13e-06) 1.0538e-04(9.13e-06)
#>   11 3.8768e-05(3.36e-06) 3.8768e-05(3.36e-06) 3.8768e-05(3.36e-06)
#>   12 1.4262e-05(1.24e-06) 1.4262e-05(1.24e-06) 1.4262e-05(1.24e-06)
#>   13 5.2466e-06(4.55e-07) 5.2466e-06(4.55e-07) 5.2466e-06(4.55e-07)
#>   14 1.9301e-06(1.67e-07) 1.9301e-06(1.67e-07) 1.9301e-06(1.67e-07)
#>   15 7.1005e-07(6.16e-08) 7.1005e-07(6.16e-08) 7.1005e-07(6.16e-08)
#>     year
#> age  2003                
#>   0  2.3212e+00(2.01e-01)
#>   1  8.5392e-01(7.40e-02)
#>   2  3.1414e-01(2.72e-02)
#>   3  1.1556e-01(1.00e-02)
#>   4  4.2514e-02(3.69e-03)
#>   5  1.5640e-02(1.36e-03)
#>   6  5.7536e-03(4.99e-04)
#>   7  2.1166e-03(1.83e-04)
#>   8  7.7867e-04(6.75e-05)
#>   9  2.8646e-04(2.48e-05)
#>   10 1.0538e-04(9.13e-06)
#>   11 3.8768e-05(3.36e-06)
#>   12 1.4262e-05(1.24e-06)
#>   13 5.2466e-06(4.55e-07)
#>   14 1.9301e-06(1.67e-07)
#>   15 7.1005e-07(6.16e-08)
#> 
#> units:  NA 

# Generate replicates based on iteration-specific multivariate distributions
# (as defined by params() and vcov())
  params(mod2)

#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>             k             t 
#>  0.45(0.0741) 10.50(0.7413) 
#> units:  NA NA 

  vcov(mod2)

#> , , 1
#> 
#>       [,1]  [,2]
#> [1,] 0.004 0.000
#> [2,] 0.000 0.001
#> 
#> , , 2
#> 
#>       [,1]  [,2]
#> [1,] 0.006 0.000
#> [2,] 0.000 0.003
#> 

  m1<-mvrnorm(mod2)
  c(params(m1))

#> [1]  0.4205440 10.0062537  0.3907853 10.9383459

# Generate replicates based on a single multivariate distribution (here the
# median of params() and vcov() is used)
  mvrnorm(2,mod2)

#> An object of class "FLModelSim"
#> Slot "model":
#> ~k^0.66 * t^0.57
#> <environment: 0x10449088>
#> 
#> Slot "params":
#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>               k               t 
#>  0.5108(0.0488) 10.5426(0.0242) 
#> units:  NA NA 
#> 
#> Slot "vcov":
#> numeric(0)
#> 
#> Slot "distr":
#> [1] "norm"
#> 

  m2<-mvrnorm(2,mod2)
  c(params(m2))

#> [1]  0.4600567 10.5227698  0.4421610 10.5547923