natural mortality

Method to simulate multivariate normal parameters for an a4aM object.

# S4 method for numeric,a4aM,missing,missing,missing,missing
mvrnorm(n = 1,
  mu)

Arguments

n
the number of iterations to be generated

mu
an a4aM object

Value

an a4aM object with n iterations

Examples

mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~k^0.66*t^0.57, params=FLPar(matrix(c(0.4,10,0.5,11),
 ncol=2, dimnames=list(params=c("k","t"), iter=1:2))),
 vcov=array(c(0.004,0.,0.,0.001,0.006,0.,0.,0.003), dim=c(2,2,2)))
mod3 <- FLModelSim(model=~1+b*v, params=FLPar(b=0.05))
mObj <- a4aM(shape=mod1, level=mod2, trend=mod3,
 range=c(min=0,max=15,minyear=2000,maxyear=2003,minmbar=0,maxmbar=0))
mObj <- mvrnorm(100, mObj)
# Generate 100 iterations with no trend over time
  m(mObj, v=c(1,1,1,1))

#> An object of class "FLQuant"
#> iters:  100 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000                 2001                 2002                
#>   0  2.3799e+00(2.47e-01) 2.3799e+00(2.47e-01) 2.3799e+00(2.47e-01)
#>   1  8.7553e-01(9.08e-02) 8.7553e-01(9.08e-02) 8.7553e-01(9.08e-02)
#>   2  3.2209e-01(3.34e-02) 3.2209e-01(3.34e-02) 3.2209e-01(3.34e-02)
#>   3  1.1849e-01(1.23e-02) 1.1849e-01(1.23e-02) 1.1849e-01(1.23e-02)
#>   4  4.3590e-02(4.52e-03) 4.3590e-02(4.52e-03) 4.3590e-02(4.52e-03)
#>   5  1.6036e-02(1.66e-03) 1.6036e-02(1.66e-03) 1.6036e-02(1.66e-03)
#>   6  5.8993e-03(6.12e-04) 5.8993e-03(6.12e-04) 5.8993e-03(6.12e-04)
#>   7  2.1702e-03(2.25e-04) 2.1702e-03(2.25e-04) 2.1702e-03(2.25e-04)
#>   8  7.9838e-04(8.28e-05) 7.9838e-04(8.28e-05) 7.9838e-04(8.28e-05)
#>   9  2.9371e-04(3.05e-05) 2.9371e-04(3.05e-05) 2.9371e-04(3.05e-05)
#>   10 1.0805e-04(1.12e-05) 1.0805e-04(1.12e-05) 1.0805e-04(1.12e-05)
#>   11 3.9749e-05(4.12e-06) 3.9749e-05(4.12e-06) 3.9749e-05(4.12e-06)
#>   12 1.4623e-05(1.52e-06) 1.4623e-05(1.52e-06) 1.4623e-05(1.52e-06)
#>   13 5.3794e-06(5.58e-07) 5.3794e-06(5.58e-07) 5.3794e-06(5.58e-07)
#>   14 1.9790e-06(2.05e-07) 1.9790e-06(2.05e-07) 1.9790e-06(2.05e-07)
#>   15 7.2803e-07(7.55e-08) 7.2803e-07(7.55e-08) 7.2803e-07(7.55e-08)
#>     year
#> age  2003                
#>   0  2.3799e+00(2.47e-01)
#>   1  8.7553e-01(9.08e-02)
#>   2  3.2209e-01(3.34e-02)
#>   3  1.1849e-01(1.23e-02)
#>   4  4.3590e-02(4.52e-03)
#>   5  1.6036e-02(1.66e-03)
#>   6  5.8993e-03(6.12e-04)
#>   7  2.1702e-03(2.25e-04)
#>   8  7.9838e-04(8.28e-05)
#>   9  2.9371e-04(3.05e-05)
#>   10 1.0805e-04(1.12e-05)
#>   11 3.9749e-05(4.12e-06)
#>   12 1.4623e-05(1.52e-06)
#>   13 5.3794e-06(5.58e-07)
#>   14 1.9790e-06(2.05e-07)
#>   15 7.2803e-07(7.55e-08)
#> 
#> units:  NA 

# Generate replicates based on iteration-specific multivariate distributions
# (as defined by params() and vcov())
  params(mod2)

#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>             k             t 
#>  0.45(0.0741) 10.50(0.7413) 
#> units:  NA NA 

  vcov(mod2)

#> , , 1
#> 
#>       [,1]  [,2]
#> [1,] 0.004 0.000
#> [2,] 0.000 0.001
#> 
#> , , 2
#> 
#>       [,1]  [,2]
#> [1,] 0.006 0.000
#> [2,] 0.000 0.003
#> 

  m1<-mvrnorm(mod2)
  c(params(m1))

#> [1]  0.4602074  9.9809938  0.4564323 10.9602921

# Generate replicates based on a single multivariate distribution (here the
# median of params() and vcov() is used)
  mvrnorm(2,mod2)

#> An object of class "FLModelSim"
#> Slot "model":
#> ~k^0.66 * t^0.57
#> <environment: 0x55cd74472360>
#> 
#> Slot "params":
#> An object of class "FLPar"
#> iters:  2 
#> 
#> params
#>                k                t 
#>  0.48863(0.0129) 10.48440(0.0353) 
#> units:  NA NA 
#> 
#> Slot "vcov":
#> numeric(0)
#> 
#> Slot "distr":
#> [1] "norm"
#> 

  m2<-mvrnorm(2,mod2)
  c(params(m2))

#> [1]  0.4379447 10.5491455  0.5258687 10.4686311