Statistical catch-at-age method

User interface to the statistical catch-at-age method of the a4a stock assessment framework.

sca(stock, indices, ...)

# S4 method for FLStock,FLIndices
sca(stock, indices, fmodel, qmodel,
  srmodel = ~factor(year), fit = "MP", mcmc = missing)

Arguments

stock

an FLStock object

indices

an FLIndices object

...

additional argument list that might never be used

fmodel

a formula object depicting the model for log fishing mortality at age

qmodel

a list of formula objects depicting the models for log survey catchability at age

srmodel

a formula object depicting the model for log recruitment

fit

character with type of fit: 'MP' or 'assessment'; the former does not require the hessian to be computed, while the latter does.

mcmc

an SCAMCMC object with the arguments to run MCMC

Value

an a4aFit or a4aFitSA object with the fit results.

Details

This method is the simple method for stock assessment. Some arguments are not accessible to make it simpler for the user. The advanced method is a4aSCA. In particular, the default for the fit argument is 'MP'. For detailed information about using the sca read the vignette 'The a4a Stock Assessment Modelling Framework' (vignette('sca')).

Examples

data(ple4)
data(ple4.index)

# fit using the default submodels
fit1 <- sca(ple4, FLIndices(ple4.index))

#> Error in get(".Generic"): object '.Generic' not found

plot(ple4 + fit1)

#> Error in plot(ple4 + fit1): object 'fit1' not found

# see default submodels (set through automated procedure)
sca(ple4, FLIndices(ple4.index), fit='assessment')

#> Error in get(".Generic"): object '.Generic' not found

# fishing mortality by age and year (separable)
fit2 <- sca(ple4, FLIndices(ple4.index), fmodel=~factor(age) + factor(year))

#> Error in get(".Generic"): object '.Generic' not found

plot(ple4 + fit2)

#> Error in plot(ple4 + fit2): object 'fit2' not found

wireframe(data~year*age, data=harvest(fit2), zlab="F")

#> Error in harvest(fit2): object 'fit2' not found

# fit2 + catcability as a smoother by age without year trend
fmod <- ~factor(age) + factor(year)
qmod <- list(~s(age, k=4))
fit3 <- sca(ple4, FLIndices(ple4.index), fmodel=fmod, qmodel=qmod)

#> Error in get(".Generic"): object '.Generic' not found

plot(ple4 + fit3)

#> Error in plot(ple4 + fit3): object 'fit3' not found

# fit3 + srmodel as a smoother by year
fmod <- ~factor(age) + factor(year)
qmod <- list(~s(age, k=4))
srmod <- ~s(year, k=45)
fit4 <- sca(ple4, FLIndices(ple4.index), fmodel=fmod, qmodel=qmod, srmodel=srmod)

#> Error in get(".Generic"): object '.Generic' not found

plot(ple4 + fit4)

#> Error in plot(ple4 + fit4): object 'fit4' not found

AIC(fit1, fit2, fit3, fit4)

#> Error in AIC(fit1, fit2, fit3, fit4): object 'fit1' not found

BIC(fit1, fit2, fit3, fit4)

#> Error in BIC(fit1, fit2, fit3, fit4): object 'fit1' not found